resize when dropping from file browser
